Add files via upload
The management of patient information is very important in healthcare. Hospitals and clinics handle a lot of data such as patient records, doctor schedules, and medical histories, etc. If this data is not stored properly, it can lead to wrong decisions, delayed actions and security risks. A well organized database can help to solve these problems.

The goal of this project is to develop a MySQL database for healthcare which can in manage the medical records in a better and more secured manner. There are still many healthcare systems that employ the use of paper based records or even complex and disorganized digital files which can lead to errors and data loss. With the right database in place, doctors and staff would be able to locate the data they require easily in which can lead to better patient care and productivity.

The main goals of this project are: It is important to organize patient’s information in a simple and straight forward manner. This is achieved by the use of unique IDs and security features to ensure that data is well protected and accurate. This will enhance the search process to an extent that doctors and staff will be able to get the records very quickly. It will assist in different departments to share data in a way that is compliant with the healthcare rules such as HIPAA.